Remove unused `helpers` package with single subpackage.
