+++ profiling/cpu.go
@@ -3,21 +3,13 @@ package profiling
 import (
 	"os"
 	"path/filepath"
-	"runtime"
 	"runtime/pprof"
 )
 
-const (
-	// CPUFilename is a filename in which the CPU profiling is stored.
-	CPUFilename = "status_cpu.prof"
-	// MemFilename is a filename in which the memory profiling is stored.
-	MemFilename = "status_mem.prof"
-)
+// CPUFilename is a filename in which the CPU profiling is stored.
+const CPUFilename = "status_cpu.prof"
 
-var (
-	cpuFile *os.File
-	memFile *os.File
-)
+var cpuFile *os.File
 
 // StartCPUProfile enables CPU profiling for the current process. While profiling,
 // the profile will be buffered and written to the file in folder dataDir.
@@ -40,18 +32,4 @@ func StopCPUProfile() error {
 	}
 	pprof.StopCPUProfile()
 	return cpuFile.Close()
-}
-
-// WriteHeapFile writes heap memory to the file.
-func WriteHeapFile(dataDir string) error {
-	if memFile == nil {
-		var err error
-		memFile, err = os.Create(filepath.Join(dataDir, MemFilename))
-		if err != nil {
-			return err
-		}
-		defer memFile.Close()
-	}
-	runtime.GC()
-	return pprof.WriteHeapProfile(memFile)
 }
